// src/models/user.ts
interface UserModel {
  username: string;
  userId: string;
  preferredUsername: string;
  name: string;
  givenName: string;
  surName: string;
  middleName: string;
  grandFatherName: string;
  email: string;
  gender: string;
  birthDateHijri: string;
  birthDate: string;
  age: string;
  nationalityName: string;
};

export default UserModel;